페어링 서버 연결

+++ app/src/main/java/com/example/whashow/data/RecommandPairList.kt
@@ -48,7 +48,7 @@ data class PairDetailResDtoByStandard(
     @SerialName("id")
     val id: Int,
     @SerialName("ratingAverage")
-    val ratingAverage: Int,
+    val ratingAverage: Float,
     @SerialName("reviewCount")
     val reviewCount: Int,
     @SerialName("title")
